+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 * Increase code coverage of "PatronNotices.js" file. Refs UICIRC-866.
 * Add "Settings (Circ): Can edit staff slips" permission. Refs UICIRC-876.
 * UI tests replacement with RTL/Jest for `notice-policy` folder in `Validation`. Refs UICIRC-811.
+* Add "requester.preferredFirstName" as staff slip token in Settings. Refs UICIRC-819.
 
 ## [7.2.1](https://github.com/folio-org/ui-circulation/tree/v7.2.1) (2022-11-29)
 [Full Changelog](https://github.com/folio-org/ui-circulation/compare/v7.2.0...v7.2.1)